Fixed Issue #1111
