+++ src/NUnitTestAdapter/ExecutionProcesses/IExecutionContext.cs
@@ -1,0 +1,15 @@
+﻿using NUnit.VisualStudio.TestAdapter.Dump;
+using NUnit.VisualStudio.TestAdapter.NUnitEngine;
+
+namespace NUnit.VisualStudio.TestAdapter.ExecutionProcesses;
+
+public interface IExecutionContext
+{
+    ITestLogger Log { get; }
+    INUnitEngineAdapter EngineAdapter { get; }
+    string TestOutputXmlFolder { get; }
+    IAdapterSettings Settings { get; }
+    IDumpXml Dump { get; }
+
+    IVsTestFilter VsTestFilter { get; }
+}
